+++ lib/money/application.ex
@@ -3,8 +3,8 @@ defmodule Money.Application do
   alias Money.ExchangeRates
   require Logger
 
-  # See http://elixir-lang.org/docs/stable/elixir/Application.html
-  # for more information on OTP Applications
+  @start_service_by_default? false
+
   def start(_type, _args) do
     import Supervisor.Spec, warn: false
 
@@ -20,12 +20,12 @@ defmodule Money.Application do
 
   # Default is to not start the exchange rate service
   defp start_exchange_rate_service? do
-    start? = Money.get_env(:exchange_rate_service, false)
+    start? = Money.get_env(:exchange_rate_service, @start_service_by_default?)
     api_module = ExchangeRates.default_config().api_module
     api_module_present? = Code.ensure_loaded?(api_module)
 
     if start? && !api_module_present? do
-      Logger.warn "ExchangeRates api module #{api_module_name(api_module)} could not be loaded. " <>
+      Logger.error "ExchangeRates api module #{api_module_name(api_module)} could not be loaded. " <>
         "  Does it exist?"
       Logger.warn "ExchangeRates service will not be started."
     end